python/requirements.txt
openai-whisper==20231117
numpy==1.23.5
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
colorama==0.4.6